move auth screens to fragment folder
