+++ app/src/main/java/ru/maxim/barybians/ui/MainActivity.kt
@@ -1,11 +1,8 @@
-package ru.maxim.barybians.ui.activity.main
+package ru.maxim.barybians.ui
 
 import android.os.Bundle
-import android.view.View
-import android.view.WindowManager
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
-import androidx.core.content.ContextCompat
 import androidx.core.view.isVisible
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
