+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="ru.maxim.barybians.ui.activity.auth.login.LoginFragment"
+        android:name="ru.maxim.barybians.ui.fragment.login.LoginFragment"
         android:label="LoginFragment"
         tools:layout="@layout/fragment_login">
